Files in folder XSSF/UserModel have been synchronized. Some methods comment out and some of tests is also. Finish them
53965 - Fixed XmlValueOutOfRangeExceptio calling getDataValidations for custom validations with XSSFSheet (poi-developers)
52255 - support adding TIFF,EPS and WPG pictures in OOXML documents (poi-developers)
52745 - fixed XSSFRichtextString.append to preserve leading / trailing spaces (poi-developers)
53374 - Avoid exceptions when parsing hyperlinks of type "javascript://"(poi-developers)
53568 - Fixed null returned by XSSFPicture.getPictureData()(poi-developers)
53950 - fixed setForceFormulaRecalculation to reset workbook-level "manual" flag(poi-developers)
53568 - Set shapes anchors in XSSF when reading from existing drawings(poi-developers)
53043 - don't duplicate hyperlink relationships when saving XSSF file(poi-developers)
51710 - fixed reading shared formulas in XSSF (poi-developers)
support setting background color of sheet tab in XSSF(poi-developers)
Bugzilla 52255 - support adding TIFF,EPS and WPG pictures in OOXML documents
52716 - tolerate hyperlinks that have neither location nor relation (poi-developers)
Validate address of hyperlinks in XSSF(poi-developers)
POIXMLPropertiesTextExtractor support for extracting custom OOXML properties as text(poi-developers)
52219 - fixed XSSFSimpleShape to set rich text attributes from XSSFRichtextString(poi-developers)
52204 - Deprecated XSSFWorkbook(String path) constructor because it does not close underlying .zip file
52268 - support cloning sheets with drawings in XSSF (poi-developers)

+++ main/SS/UserModel/Shape.cs
@@ -15,7 +15,7 @@ namespace NPOI.SS.UserModel
         int LineStyleColor { get; }
 
         int FillColor { get; set; }
-        int LineWidth { get; set; }
+        double LineWidth { get; set; }
         LineStyle LineStyle { get; set; }
         bool IsNoFill { get; set; }
         int CountOfAllChildren { get; }
